requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
Pillow>=10.2.0
mutagen==1.47.0
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
yt-dlp==2025.9.5
rembg
netifaces==0.11.0
qrcode
playwright==1.48.0
moviepy==1.0.3
